refactor: ValidatorTest 클래스 분리
